OD-10

+++ src/contracts/proxies/Vault721.sol
@@ -32,7 +32,7 @@ contract Vault721 is ERC721EnumerableUpgradeable {
   address public timelockController;
   IODSafeManager public safeManager;
   NFTRenderer public nftRenderer;
-  uint8 public blockDelay;
+  uint256 public blockDelay;
   uint256 public timeDelay;
 
   string public contractMetaData =
@@ -193,7 +193,7 @@ contract Vault721 is ERC721EnumerableUpgradeable {
   /**
    * @dev allows DAO to update the block delay
    */
-  function updateBlockDelay(uint8 _blockDelay) external onlyGovernance {
+  function updateBlockDelay(uint256 _blockDelay) external onlyGovernance {
     blockDelay = _blockDelay;
   }
